Updated the Title on the listvouchers.phpinfo
Updated the listjacks.php page so that you can click on a building in the tabs across the top and then just deal with the records for that building.
	**NOTE:	Duplicate the code on Line 90 for each building you want to add.
		Change EX to the initials of the building in both spots

Added EX.php to be used with the listjacks.php page rename and duplicate this page for each building you wish to add.
	Change the following code on the page before duplicating EX.php:
		Line 153 - Change Building Name Here to be the full name of the building
		Line 228 - Change Building Initials Here to be the initials of the building
		Line 322 - Change 00000 to be the nameid that matches the locattions table nameid Column
		Line 339 - Change 00000 to be the nameid that matches the locattions table nameid Column